requirements.txt
python-dotenv
psycopg2
requests
imgkit
atoma